Refactor student report shell components
